refactor(relay): extract ChecksumUpdate into a shared crate
The incremental checksum updates of the eBPF TURN router are useful
outside of it: connlib's NAT rewrites packet headers on every packet
of a DNS resource and can maintain the checksums with the same diffs
instead of recomputing them over the full payload. Move the
`ChecksumUpdate` abstraction into a dedicated `no_std` crate; it is
free of unbounded loops, so it remains usable from eBPF.
