+++ rust/libs/incremental-inet-checksum/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "incremental-inet-checksum"
+version = "0.1.0"
+edition = { workspace = true }
+license = { workspace = true }
+
+[lints]
+workspace = true
